refactor: UserResponseをpresentation/user/パッケージに移動
ユーザー情報の共通レスポンスDTOであるUserResponseをauthパッケージから
userパッケージに移動し、参照元のimportを更新した。

Closes #56
