+++ backend/src/main/java/com/youmorry/expensetracker/presentation/auth/AuthResponse.java
@@ -1,6 +1,7 @@
 package com.youmorry.expensetracker.presentation.auth;
 
 import com.youmorry.expensetracker.application.AuthService.AuthResult;
+import com.youmorry.expensetracker.presentation.user.UserResponse;
 
 /** Google 認証レスポンス。JWT アクセストークンとユーザー情報を返す。 */
 public record AuthResponse(String accessToken, UserResponse user) {
